source 수정

+++ src/main/java/com/tsp/new_tsp_admin/api/model/service/AdminModelJpaService.java
@@ -21,7 +21,7 @@ public interface AdminModelJpaService {
      * </pre>
      *
      */
-    Long findModelsCount(Map<String, Object> modelMap) throws Exception;
+    int findModelsCount(Map<String, Object> modelMap) throws Exception;
 
     /**
      * <pre>
